Changes to support Llama3.1

+++ generate_act_scale_shift.py
@@ -97,7 +97,7 @@ def get_act_shifts(model, dataloader, num_samples=128):
 
 
 def build_model_and_tokenizer(model_name):
-    kwargs = {"torch_dtype": torch.float16, "device_map": "auto"}
+    kwargs = {"torch_dtype": torch.bfloat16, "device_map": "auto"}
     tokenizer = AutoTokenizer.from_pretrained(model_name)
     model = AutoModelForCausalLM.from_pretrained(model_name, **kwargs)
     return model, tokenizer
